+++ tests/test_main.py
@@ -43,7 +43,9 @@ def test_without_passed_user_and_passwd(monkeypatch):
             fast_mode=False,
             dry_run=False,
             encoding='',
-            skip_dialect=False
+            skip_dialect=False,
+            skip_row_id=False,
+            output_delimiter=None
         )
 
 
@@ -85,7 +87,9 @@ def test_keep_cols(monkeypatch):
             fast_mode=False,
             dry_run=False,
             encoding='',
-            skip_dialect=False
+            skip_dialect=False,
+            skip_row_id=False,
+            output_delimiter=None
         )
 
 
@@ -171,7 +175,9 @@ def test_datarobot_key(monkeypatch):
             auto_sample=True,
             dry_run=False,
             encoding='',
-            skip_dialect=False
+            skip_dialect=False,
+            skip_row_id=False,
+            output_delimiter=None
         )
 
 
@@ -214,7 +220,9 @@ def test_encoding_options(monkeypatch):
             auto_sample=True,
             dry_run=False,
             encoding='utf-8',
-            skip_dialect=True
+            skip_dialect=True,
+            skip_row_id=False,
+            output_delimiter=None
         )
 
 
@@ -255,3 +263,93 @@ def test_no_required_params(monkeypatch):
         with pytest.raises(SystemExit):
             main(argv=main_args)
     assert not mock_method.called
+
+
+def test_output_delimiter(monkeypatch):
+    main_args = ['--host',
+                 'http://localhost:53646/api',
+                 '56dd9570018e213242dfa93c',
+                 '56dd9570018e213242dfa93d',
+                 'tests/fixtures/temperatura_predict.csv',
+                 '--output_delimiter=tab',
+                 '--encoding=utf-8', '--skip_dialect']
+
+    monkeypatch.setattr('datarobot_batch_scoring.main.UI', mock.Mock(spec=UI))
+
+    with mock.patch(
+            'datarobot_batch_scoring.main'
+            '.run_batch_predictions') as mock_method:
+        main(argv=main_args)
+        mock_method.assert_called_once_with(
+            base_url='http://localhost:53646/api/v1/',
+            base_headers={},
+            user=mock.ANY,
+            pwd=None,
+            api_token=None,
+            create_api_token=False,
+            pid='56dd9570018e213242dfa93c',
+            lid='56dd9570018e213242dfa93d',
+            n_retry=3,
+            concurrent=4,
+            resume=False,
+            n_samples=False,
+            out_file='out.csv',
+            keep_cols=None,
+            delimiter=None,
+            dataset='tests/fixtures/temperatura_predict.csv',
+            pred_name=None,
+            timeout=30,
+            ui=mock.ANY,
+            fast_mode=False,
+            auto_sample=True,
+            dry_run=False,
+            encoding='utf-8',
+            skip_dialect=True,
+            skip_row_id=False,
+            output_delimiter='\t'
+        )
+
+
+def test_skip_row_id(monkeypatch):
+    main_args = ['--host',
+                 'http://localhost:53646/api',
+                 '56dd9570018e213242dfa93c',
+                 '56dd9570018e213242dfa93d',
+                 'tests/fixtures/temperatura_predict.csv',
+                 '--skip_row_id',
+                 '--encoding=utf-8', '--skip_dialect']
+
+    monkeypatch.setattr('datarobot_batch_scoring.main.UI', mock.Mock(spec=UI))
+
+    with mock.patch(
+            'datarobot_batch_scoring.main'
+            '.run_batch_predictions') as mock_method:
+        main(argv=main_args)
+        mock_method.assert_called_once_with(
+            base_url='http://localhost:53646/api/v1/',
+            base_headers={},
+            user=mock.ANY,
+            pwd=None,
+            api_token=None,
+            create_api_token=False,
+            pid='56dd9570018e213242dfa93c',
+            lid='56dd9570018e213242dfa93d',
+            n_retry=3,
+            concurrent=4,
+            resume=False,
+            n_samples=False,
+            out_file='out.csv',
+            keep_cols=None,
+            delimiter=None,
+            dataset='tests/fixtures/temperatura_predict.csv',
+            pred_name=None,
+            timeout=30,
+            ui=mock.ANY,
+            fast_mode=False,
+            auto_sample=True,
+            dry_run=False,
+            encoding='utf-8',
+            skip_dialect=True,
+            skip_row_id=True,
+            output_delimiter=None
+        )